+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/ast/impl/SuppressionCommentImpl.java
@@ -15,7 +15,8 @@ import net.sourceforge.pmd.reporting.ViolationSuppressor.SuppressionCommentWrapp
  *
  * @param <T> Type of Reportable (node, token, lambda)
  *
- * @experimental Since 7.14.0. See <a href="https://github.com/pmd/pmd/pull/5609">[core] Add rule to report unnecessary suppression comments/annotations #5609</a>
+ * @since 7.14.0
+ * @experimental See <a href="https://github.com/pmd/pmd/pull/5609">[core] Add rule to report unnecessary suppression comments/annotations #5609</a>
  */
 @Experimental
 public class SuppressionCommentImpl<T extends Reportable> implements SuppressionCommentWrapper {