+++ pmd-visualforce/src/main/java/net/sourceforge/pmd/lang/visualforce/ast/VfTypedNode.java
@@ -13,7 +13,7 @@ public interface VfTypedNode extends VfNode {
 
     /**
      * Returns the data type this node refers to. A null value indicates that no matching Metadata was found for this
-     * node. null differs from {@link DataType#Unknown} which indicates that Metadata was found but it wasn't mappable
+     * node. null differs from {@link DataType#UNKNOWN} which indicates that Metadata was found but it wasn't mappable
      * to one of the enums.
      *
      * <p>Example XPath 1.0 and 2.0: {@code //Identifier[@DataType='DateTime']}