+++ docs/pages/release_notes.md
@@ -24,11 +24,45 @@ This is a {{ site.pmd.release_type }} release.
 
 ### 🚀️ New and noteworthy
 
+### 🌟️ New and Changed Rules
+#### New Rules
+* The new Java rule {% rule java/codestyle/UnnecessaryInterfaceDeclaration %} detects classes that
+  implement interfaces that are already implemented by its superclass, and interfaces
+  that extend other interfaces already declared by their superinterfaces.  
+  These declarations are redundant and can be removed to simplify the code.
+
+#### Changed Rules
+* The rule {% rule java/errorprone/CloseResource %} introduces a new property, `allowedResourceMethodPatterns`,
+  which lets you specify method invocation patterns whose return values are resources managed externally.
+  This is useful for ignoring managed resources - for example, `Reader`/`Writer` instances obtained from
+  `HttpServletRequest`/`HttpServletResponse` - because the servlet container, not application code,
+  is responsible for closing them. By default, the rule ignores `InputStream`/`OutputStream`/`Reader`/`Writer`
+  resources returned by methods on `(Http)ServletRequest` and `(Http)ServletResponse`
+  (both `javax.servlet` and `jakarta.servlet`).
+
 ### 🐛️ Fixed Issues
+* doc
+  * [#6396](https://github.com/pmd/pmd/pull/6396): \[doc] Mention test-pmd-tool as alternative for testing
+* java-bestpractices
+  * [#6431](https://github.com/pmd/pmd/issues/6431): \[java] UnitTestShouldIncludeAssert: False positive with SoftAssertionsExtension on parent/grandparent classes
+* java-codestyle
+  * [#6458](https://github.com/pmd/pmd/pull/6458): \[java] New Rule: UnnecessaryInterfaceDeclaration
 * java-errorprone
   * [#5787](https://github.com/pmd/pmd/issues/5787): \[java] InvalidLogMessageFormat: false positive with lombok @Value generated methods
+  * [#6436](https://github.com/pmd/pmd/issues/6436): \[java] CloseResource: Allow to ignore managed resources
 
 ### 🚨️ API Changes
+
+#### Deprecations
+* core
+  * {%jdoc core::renderers.CodeClimateIssue %}: This class is an implementation detail of
+    {%jdoc core::renderers.CodeClimateRenderer %}. It will be internalized in a future release.
+* visualforce
+  * {%jdoc visualforce::lang.visualforce.DataType %}. The enum constants have been renamed to follow Java naming
+    conventions. The old enum constants are deprecated and should no longer be used.  
+    The method {%jdoc !!visualforce::lang.visualforce.DataType#fromString(java.lang.String) %} will return the new
+    enum constants.  
+    Use {%jdoc !!visualforce::lang.visualforce.DataType#fieldTypeNameOf() %} to get the original field type name.
 
 ### ✨️ Merged pull requests
 <!-- content will be automatically generated, see /do-release.sh -->
@@ -40,4 +74,3 @@ This is a {{ site.pmd.release_type }} release.
 <!-- content will be automatically generated, see /do-release.sh -->
 
 {% endtocmaker %}
-